+++ fluentd/lib/puppet/parser/functions/fluent_config.rb
@@ -1,29 +1,43 @@
+# This file must be compatible with Ruby 1.8.7 in order to work on EL6.
 module Puppet::Parser::Functions
-  newfunction(:fluent_plugin_config, type: :rvalue) do |args|
+
+  # Generate fluentd config from Hash.
+  newfunction(:fluent_config, :type => :rvalue) do |args|
+    config = args[0]
+
+    header = "# Managed by Puppet.\n"
+
+    # NOTE: Hash iteration order is arbitrary in ruby 1.8.7
+    # https://projects.puppetlabs.com/issues/16266
+    config.keys.sort.inject(header) do |result, plugin_type|
+      plugin_config = config[plugin_type]
+      result << function_fluent_plugin_config([plugin_type, plugin_config])
+    end.chomp
+  end
+
+  # Generate fluentd plugin config from Hash
+  newfunction(:fluent_plugin_config, :type => :rvalue) do |args|
     plugin_type = args[0]
     plugin_config = args[1]
 
     tag_pattern = plugin_config.delete('tag_pattern')
 
-    config_body = plugin_config.each_with_object('') do |(key, value), result|
+    config_body = ''
+
+    # NOTE: Hash iteration order is arbitrary in ruby 1.8.7
+    # https://projects.puppetlabs.com/issues/16266
+    plugin_config.keys.sort.each do |key|
+      value = plugin_config[key]
+
       if value.is_a?(Array)
         value.each do |plugin_sub_config|
-          result << function_fluent_plugin_config([key, plugin_sub_config])
+          config_body << function_fluent_plugin_config([key, plugin_sub_config])
         end
       else
-        result << [key, value].join(' ') << "\n"
+        config_body << [key, value].join(' ') << "\n"
       end
     end
 
     "<#{plugin_type} #{tag_pattern}>\n#{config_body}</#{plugin_type}>\n\n"
-  end
-
-  newfunction(:fluent_config, type: :rvalue) do |args|
-    config = args[0]
-
-    header = "# Managed by Puppet.\n"
-    config.each_with_object(header) do |(plugin_type, plugin_config), result|
-      result << function_fluent_plugin_config([plugin_type, plugin_config])
-    end.chomp
   end
 end