+++ fluentd/spec/classes/service_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe 'fluentd::service' do
   context 'on RedHat based system' do
     let(:facts) { { osfamily: 'RedHat' } }
 
-    it { is_expected.to contain_service('td-agent') }
+    it { is_expected.to contain_service('td-agent').with(provider: 'redhat') }
   end
 
   context 'on Debian based system' do
@@ -12,6 +12,6 @@ RSpec.describe 'fluentd::service' do
       { osfamily: 'Debian', lsbdistid: 'Ubuntu', lsbdistcodename: 'trusty' }
     end
 
-    it { is_expected.to contain_service('td-agent') }
+    it { is_expected.to contain_service('td-agent').without(:provider) }
   end
 end