+++ fluentd/spec/spec_helper_acceptance.rb
@@ -12,8 +12,8 @@ RSpec.configure do |c|
     puppet_module_install(source: module_root, module_name: 'fluentd')
 
     hosts.each do |host|
-      on host, puppet('module', 'install', 'puppetlabs-stdlib'), acceptable_exit_codes: [0, 1]
-      on host, puppet('module', 'install', 'puppetlabs-apt'), acceptable_exit_codes: [0, 1]
+      on host, puppet('module', 'install', 'puppetlabs-stdlib')
+      on host, puppet('module', 'install', 'puppetlabs-apt')
     end
   end
 end