+++ opendaylight/spec/acceptance/class_spec.rb
@@ -91,4 +91,77 @@ describe 'opendaylight class' do
       port_config_validations(odl_rest_port: 7777)
     end
   end
+
+  describe 'testing custom logging verbosity' do
+    context 'using default log levels' do
+      # Call specialized helper fn to install OpenDaylight
+      install_odl
+
+      # Call specialized helper fn for custom logger verbosity validations
+      log_level_validations
+    end
+
+    context 'adding one custom log level' do
+      custom_log_levels = { 'org.opendaylight.ovsdb' => 'TRACE' }
+
+      # Call specialized helper fn to install OpenDaylight
+      install_odl(log_levels: custom_log_levels)
+
+      # Call specialized helper fn for custom logger verbosity validations
+      log_level_validations(log_levels: custom_log_levels)
+    end
+
+    context 'adding two custom log level' do
+      custom_log_levels = { 'org.opendaylight.ovsdb' => 'TRACE',
+                            'org.opendaylight.ovsdb.lib' => 'INFO' }
+
+      # Call specialized helper fn to install OpenDaylight
+      install_odl(log_levels: custom_log_levels)
+
+      # Call specialized helper fn for custom logger verbosity validations
+      log_level_validations(log_levels: custom_log_levels)
+    end
+  end
+
+  describe 'testing ODL OVSDB L3 config' do
+    context 'using enable_l3 default' do
+      # Call specialized helper fn to install OpenDaylight
+      install_odl
+
+      # Call specialized helper fn for ODL OVSDB L3 config validations
+      enable_l3_validations
+    end
+
+    context 'using "no" for enable_l3' do
+      # Call specialized helper fn to install OpenDaylight
+      install_odl(enable_l3: 'no')
+
+      # Call specialized helper fn for ODL OVSDB L3 config validations
+      enable_l3_validations(enable_l3: 'no')
+    end
+
+    context 'using "yes" for enable_l3' do
+      # Call specialized helper fn to install OpenDaylight
+      install_odl(enable_l3: 'yes')
+
+      # Call specialized helper fn for ODL OVSDB L3 config validations
+      enable_l3_validations(enable_l3: 'yes')
+    end
+
+    context 'using false for enable_l3' do
+      # Call specialized helper fn to install OpenDaylight
+      install_odl(enable_l3: false)
+
+      # Call specialized helper fn for ODL OVSDB L3 config validations
+      enable_l3_validations(enable_l3: false)
+    end
+
+    context 'using true for enable_l3' do
+      # Call specialized helper fn to install OpenDaylight
+      install_odl(enable_l3: true)
+
+      # Call specialized helper fn for ODL OVSDB L3 config validations
+      enable_l3_validations(enable_l3: true)
+    end
+  end
 end