+++ opendaylight/spec/classes/opendaylight_spec.rb
@@ -11,7 +11,7 @@ describe 'opendaylight' do
         operatingsystem = 'Fedora'
 
         # All tests for supported versions of Fedora
-        ['20', '21'].each do |operatingsystemmajrelease|
+        ['22', '23'].each do |operatingsystemmajrelease|
           context "#{operatingsystemmajrelease}" do
             let(:facts) {{
               :osfamily => osfamily,
@@ -22,20 +22,35 @@ describe 'opendaylight' do
             # Note that this function is defined in spec_helper
             generic_tests
 
-            # Run test that specialize in checking rpm-based installs
+            # Run tests that specialize in checking rpm-based installs
             # NB: Only testing defaults here, specialized rpm tests elsewhere
             # Note that this function is defined in spec_helper
             rpm_install_tests(operatingsystem: operatingsystem)
 
-            # Run test that specialize in checking Karaf feature installs
+            # Run tests that specialize in checking Karaf feature installs
             # NB: Only testing defaults here, specialized Karaf tests elsewhere
             # Note that this function is defined in spec_helper
             karaf_feature_tests
+
+            # Run tests that specialize in checking ODL's REST port config
+            # NB: Only testing defaults here, specialized log level tests elsewhere
+            # Note that this function is defined in spec_helper
+            odl_rest_port_tests
+
+            # Run tests that specialize in checking custom log level config
+            # NB: Only testing defaults here, specialized log level tests elsewhere
+            # Note that this function is defined in spec_helper
+            log_level_tests
+
+            # Run tests that specialize in checking ODL OVSDB L3 config
+            # NB: Only testing defaults here, specialized enabling L3 tests elsewhere
+            # Note that this function is defined in spec_helper
+            enable_l3_tests
           end
         end
 
         # All tests for unsupported versions of Fedora
-        ['19'].each do |operatingsystemmajrelease|
+        ['21'].each do |operatingsystemmajrelease|
           context "#{operatingsystemmajrelease}" do
             let(:facts) {{
               :osfamily => osfamily,
@@ -75,11 +90,26 @@ describe 'opendaylight' do
             # NB: Only testing defaults here, specialized Karaf tests elsewhere
             # Note that this function is defined in spec_helper
             karaf_feature_tests
+
+            # Run tests that specialize in checking ODL's REST port config
+            # NB: Only testing defaults here, specialized log level tests elsewhere
+            # Note that this function is defined in spec_helper
+            odl_rest_port_tests
+
+            # Run test that specialize in checking custom log level config
+            # NB: Only testing defaults here, specialized log level tests elsewhere
+            # Note that this function is defined in spec_helper
+            log_level_tests
+
+            # Run tests that specialize in checking ODL OVSDB L3 config
+            # NB: Only testing defaults here, specialized enabling L3 tests elsewhere
+            # Note that this function is defined in spec_helper
+            enable_l3_tests
           end
         end
 
         # All tests for unsupported versions of CentOS
-        ['5', '6', '8'].each do |operatingsystemmajrelease|
+        ['6'].each do |operatingsystemmajrelease|
           context "#{operatingsystemmajrelease}" do
             let(:facts) {{
               :osfamily => osfamily,
@@ -135,11 +165,26 @@ describe 'opendaylight' do
             # NB: Only testing defaults here, specialized Karaf tests elsewhere
             # Note that this function is defined in spec_helper
             karaf_feature_tests
+
+            # Run tests that specialize in checking ODL's REST port config
+            # NB: Only testing defaults here, specialized log level tests elsewhere
+            # Note that this function is defined in spec_helper
+            odl_rest_port_tests
+
+            # Run test that specialize in checking custom log level config
+            # NB: Only testing defaults here, specialized log level tests elsewhere
+            # Note that this function is defined in spec_helper
+            log_level_tests
+
+            # Run tests that specialize in checking ODL OVSDB L3 config
+            # NB: Only testing defaults here, specialized enabling L3 tests elsewhere
+            # Note that this function is defined in spec_helper
+            enable_l3_tests
           end
         end
 
         # All tests for unsupported versions of Ubuntu
-        ['12.04', '12.10', '13.04', '13.10', '14.10', '15.04'].each do |operatingsystemmajrelease|
+        ['12.04', '15.10'].each do |operatingsystemmajrelease|
           context "#{operatingsystemmajrelease}" do
             let(:facts) {{
               :osfamily => osfamily,
@@ -311,6 +356,185 @@ describe 'opendaylight' do
       # Run test that specialize in checking ODL REST port config
       # Note that this function is defined in spec_helper
       odl_rest_port_tests(odl_rest_port: 7777)
+    end
+  end
+
+  # All custom log level tests
+  describe 'custom log level tests' do
+    # Non-OS-type tests assume CentOS 7
+    #   See issue #43 for reasoning:
+    #   https://github.com/dfarrell07/puppet-opendaylight/issues/43#issue-57343159
+    osfamily = 'RedHat'
+    operatingsystem = 'CentOS'
+    operatingsystemmajrelease = '7'
+    context 'using default log levels' do
+      let(:facts) {{
+        :osfamily => osfamily,
+        :operatingsystem => operatingsystem,
+        :operatingsystemmajrelease => operatingsystemmajrelease,
+      }}
+
+      let(:params) {{ }}
+
+      # Run shared tests applicable to all supported OSs
+      # Note that this function is defined in spec_helper
+      generic_tests
+
+      # Run test that specialize in checking custom log level config
+      # Note that this function is defined in spec_helper
+      log_level_tests
+    end
+
+    context 'adding one custom log level' do
+      let(:facts) {{
+        :osfamily => osfamily,
+        :operatingsystem => operatingsystem,
+        :operatingsystemmajrelease => operatingsystemmajrelease,
+      }}
+
+      custom_log_levels = { 'org.opendaylight.ovsdb' => 'TRACE' }
+
+      let(:params) {{
+        :log_levels => custom_log_levels,
+      }}
+
+      # Run shared tests applicable to all supported OSs
+      # Note that this function is defined in spec_helper
+      generic_tests
+
+      # Run test that specialize in checking log level config
+      # Note that this function is defined in spec_helper
+      log_level_tests(log_levels: custom_log_levels)
+    end
+
+    context 'adding two custom log levels' do
+      let(:facts) {{
+        :osfamily => osfamily,
+        :operatingsystem => operatingsystem,
+        :operatingsystemmajrelease => operatingsystemmajrelease,
+      }}
+
+      custom_log_levels = { 'org.opendaylight.ovsdb' => 'TRACE',
+                         'org.opendaylight.ovsdb.lib' => 'INFO' }
+
+      let(:params) {{
+        :log_levels => custom_log_levels,
+      }}
+
+      # Run shared tests applicable to all supported OSs
+      # Note that this function is defined in spec_helper
+      generic_tests
+
+      # Run test that specialize in checking log level config
+      # Note that this function is defined in spec_helper
+      log_level_tests(log_levels: custom_log_levels)
+    end
+  end
+
+  # All OVSDB L3 enable/disable tests
+  describe 'OVSDB L3 enable/disable tests' do
+    # Non-OS-type tests assume CentOS 7
+    #   See issue #43 for reasoning:
+    #   https://github.com/dfarrell07/puppet-opendaylight/issues/43#issue-57343159
+    osfamily = 'RedHat'
+    operatingsystem = 'CentOS'
+    operatingsystemmajrelease = '7'
+    context 'using enable_l3 default' do
+      let(:facts) {{
+        :osfamily => osfamily,
+        :operatingsystem => operatingsystem,
+        :operatingsystemmajrelease => operatingsystemmajrelease,
+      }}
+
+      let(:params) {{ }}
+
+      # Run shared tests applicable to all supported OSs
+      # Note that this function is defined in spec_helper
+      generic_tests
+
+      # Run test that specialize in checking ODL OVSDB L3 config
+      # Note that this function is defined in spec_helper
+      enable_l3_tests
+    end
+
+    context 'using "no" for enable_l3' do
+      let(:facts) {{
+        :osfamily => osfamily,
+        :operatingsystem => operatingsystem,
+        :operatingsystemmajrelease => operatingsystemmajrelease,
+      }}
+
+      let(:params) {{
+        :enable_l3 => 'no',
+      }}
+
+      # Run shared tests applicable to all supported OSs
+      # Note that this function is defined in spec_helper
+      generic_tests
+
+      # Run test that specialize in checking ODL OVSDB L3 config
+      # Note that this function is defined in spec_helper
+      enable_l3_tests(enable_l3: 'no')
+    end
+
+    context 'using "yes" for enable_l3' do
+      let(:facts) {{
+        :osfamily => osfamily,
+        :operatingsystem => operatingsystem,
+        :operatingsystemmajrelease => operatingsystemmajrelease,
+      }}
+
+      let(:params) {{
+        :enable_l3 => 'yes',
+      }}
+
+      # Run shared tests applicable to all supported OSs
+      # Note that this function is defined in spec_helper
+      generic_tests
+
+      # Run test that specialize in checking ODL OVSDB L3 config
+      # Note that this function is defined in spec_helper
+      enable_l3_tests(enable_l3: 'yes')
+    end
+
+    context 'using false for enable_l3' do
+      let(:facts) {{
+        :osfamily => osfamily,
+        :operatingsystem => operatingsystem,
+        :operatingsystemmajrelease => operatingsystemmajrelease,
+      }}
+
+      let(:params) {{
+        :enable_l3 => false ,
+      }}
+
+      # Run shared tests applicable to all supported OSs
+      # Note that this function is defined in spec_helper
+      generic_tests
+
+      # Run test that specialize in checking ODL OVSDB L3 config
+      # Note that this function is defined in spec_helper
+      enable_l3_tests(enable_l3: false)
+    end
+
+    context 'using true for enable_l3' do
+      let(:facts) {{
+        :osfamily => osfamily,
+        :operatingsystem => operatingsystem,
+        :operatingsystemmajrelease => operatingsystemmajrelease,
+      }}
+
+      let(:params) {{
+        :enable_l3 => true,
+      }}
+
+      # Run shared tests applicable to all supported OSs
+      # Note that this function is defined in spec_helper
+      generic_tests
+
+      # Run test that specialize in checking ODL OVSDB L3 config
+      # Note that this function is defined in spec_helper
+      enable_l3_tests(enable_l3: true)
     end
   end