+++ fluentd/CHANGELOG.md
@@ -1,4 +1,12 @@
-## Unreleased
+## 2016-02-03 - Release v. 0.4.0
+
+ - Support CentOS 6
+
+## 2016-01-22 - Release v. 0.3.2
+
+ - Purge unmanaged config files
+ - Manage td-agent.conf file with a fully qualified path
+ - Fix the issue with td-agent service being enabled on each run on EL7
 
 ## 2015-12-02 - Release v. 0.3.1